Move popvers to other files, simplify css
