+++ resources/tentative/complex-static-html/src/big-dom-generator/components/analytics-pop-over.jsx
@@ -1,0 +1,22 @@
+import { VerticalPopOver } from "./pop-over.jsx";
+
+const ProgressBar = ({ teamNumber, value }) => {
+  return (
+      <div className="ui spectrum-ProgressBar spectrum-Meter--sizeS is-negative analytics-bar" value={value} role="progressbar" aria-valuenow={value} aria-valuemin="0" aria-valuemax="100">
+          <div className="ui spectrum-FieldLabel spectrum-FieldLabel--sizeS spectrum-ProgressBar-label">Team {teamNumber} progress</div>
+          <div className="ui spectrum-FieldLabel spectrum-FieldLabel--sizeS spectrum-ProgressBar-percentage">{value}%</div>
+          <div className="ui spectrum-ProgressBar-track">
+              <div className="ui spectrum-ProgressBar-fill" style={{ width: value.toString() + "%" }}></div>
+          </div>
+      </div>
+  );
+};
+
+export const AnalyticsPopOver = () => {
+  const children = [];
+  const numProgressBars = 5;
+  for (let i = 0; i < numProgressBars; i++) {
+      children.push(<ProgressBar key={i} teamNumber={i} value={((i + 1) * 10) % 100} />);
+  }
+  return <VerticalPopOver>{children}</VerticalPopOver>;
+};
